% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_atlas}
\alias{load_atlas}
\title{Load atlas data from a directory of CSV files}
\usage{
load_atlas(path)
}
\arguments{
\item{path}{Directory containing the atlas CSVs.}
}
\value{
A list with elements \code{obs}, \code{markers}, \code{diet_markers}, \code{genes}, \code{var}.
}
\description{
Load atlas data from a directory of CSV files
}
